requirements.txt
blurhash==1.1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
geographiclib==1.50
geopy==1.20.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
Pillow==7.0.0
pkg-resources==0.0.0
requests==2.22.0
urllib3==1.25.7
Werkzeug==0.16.0
